created an OptionStore for options management in flux
