+++ src/lib/WingTron.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import OptionsStore from "./optionsStore.js";
 import ApplicationInterface from "./applicationInterface.js";
 
 // provide another component for the scoreboard & control panel rendering and api call to the start game function
@@ -6,18 +7,7 @@ class WingTron extends Component {
   constructor(props) {
     super(props)
     this.state = {
-      gameOptions: {
-        fieldWidth: "300", fieldHeight: "300", matches: "10",
-      },
-      playerOptions: [
-        {"startCoord": [150,150], "direction": 0, "keyCodes": {"leftCode": 37, "rightCode": 39}, "colour": "rgba(255,0,0,0.5)",},
-        {"startCoord": [150,150], "direction": Math.PI, "keyCodes": {"leftCode": 65, "rightCode": 68}, "colour": "rgba(0,0,255,0.5)",},
-      ],
-      environmentOptions: {
-        scoreboardFunction: function (a) {
-          console.log(a);
-        },
-      },
+      options: OptionsStore.getAll(),
     };
     // set a load of presets incase no control panel is provided
     // also render a "click here to start" text on the canvas if there is no control panel
@@ -28,11 +18,11 @@ class WingTron extends Component {
   }
   updateCanvas() {
     var canvas = this.refs.canvas;
-    canvas.width = this.state.gameOptions.fieldWidth;
-    canvas.height = this.state.gameOptions.fieldHeight;
+    canvas.width = this.state.options.gameOptions.fieldWidth;
+    canvas.height = this.state.options.gameOptions.fieldHeight;
     canvas.style.width = '100%';
 
-    ApplicationInterface.startGame(this.state, canvas);
+    ApplicationInterface.startGame(this.state.options, canvas);
   }
   render() {
     const { color, children } = this.props;
